Track play state with a flux-style store.
This implementation may not survive long-term, but this was a useful
exercise to see how different state management tactics work with Yew.

Some observations at this early stage:

- Flux is intended to make data available across a range of components
  without requiring those components to feed the data to each other
  (decoupling data from the tree).
- The app is basically one big component right now, so this feels heavy.
- Store changes are broadcast as one big (fat) "changed" signal, and
  subscribers have to filter/cherry-pick the meaning from those signals.
- A Yew "store" is closer to a single reducer in redux.
- Store subscribers need to capture copies of store data on their own
  fields to be able to render stuff. This is similar to the pre-hooks "map
  state to props" workflow.

Ultimately, much of the app state will be offloaded to a server and
sync'd via websocket. This store design may or may not turn out to serve
well in this situation.