Add PlayerCards component to display player selection states.
Originally the plan for this was to not send the component any props at
all, instead it would subscribe to the store.

Unfortunately, it seems like only the first bridge opened (currently
this is in the root app component) gets the special "initial update"
message, so this essentially means you have to either artificially cause
the component to update after the initial mount, or you have to
artificially cause the store to change after the component has mounted.
Neither seem like great choices.

For now, rather than ripping the store out of the app (since it seems to
have only limited usefulness as-is), the root reads from the store and
passes the data down to child components as needed. This may prove
useful in the long term since deeper components can still dispatch
messages to the store and cause the root to receive new data.

When we get into the whole websocket part of this project, we can
exploit this to sync data between clients.