Add WS server, and use server state instead of an app-local store.
Pass `Addr<GameState>` around so `PhiSocket`s can push messages to it.

Add actors on backend, move common types, refactor client.

Continuing to stub things out. The actors added to the backend are yet
to do anything, but some of the messaging pathways are framed by the
message types plus the addition of `Addr<A>` types to various spots.

Data types that are mostly common between client and server have been
moved to the common crate with small tweaks as needed to make the
division between client/server more obvious. We don't want to leak
either client-specific or server-specific data on either side.

Create random admin key on app startup.

Reorganize "commands" in terms of client vs server.

The client commands have been moved over to the common crate so they can
be shared with both the client and server.

The server commands are mainly going to be the concern of socket
handlers when they start/stop to manage some state controlled by the
PlaySession, so there's no need for these to appear on the client-side.

Delegate commands from `PhiSocket` to `PlaySession`.

The socket handler at this point does very little. It'll parse json
commands from the text it receives over the socket, then immediately
delegate to the PlaySession actor.

There are a couple special cases, like the initial registration of a
player, and the closing of a socket connection.

Add a way to notify all connections about changes to `GameState`.

Implements a method on `PlaySession` that feeds a clone of the current
`GameState` to each open websocket.

Also adds messaging types/plumbing to pass that changed game state all
the way down to the socket.

Add implementation for `AddConnection`/`RemoveConnection`.

Remove store. Adopt server state over WS.