Fold ConnectionId/PlayerId into the same. 1 player = 1 connection.
Simplifying the backend storage for sockets such that a player is the
same as a connection (thus removing the ability to have a single player
with more than one open socket).

Additionally, move the registration process so that it is managed by the
server instead of having clients identify themselves.

The new flow is like:

- client connects
- server ack's and pushes a client id back to the client