Apollo client schema dl/codegen ignore server if local file specified.
I had thought the schema download would ignore a local file if both the
url and local file were specified, but it turns out "no."

If we want to avoid contacting a running backend for apollo-related
tasks in CI or even during development, we'll need a 2nd apollo config
tht specifies the local schema file instead of the url.

It's fine I guess.

For now, always look at the backend when doing codegen+schema downloads.