chore: update client deps
Dependencies that come from CRA, apollo, tailwind, and so on have been
bumped to the latest possible versions.

Note that with apollo, there are issues related to peer dependencies on
`graphql`, so we're pinned to 15.x rather than 16.x. It's some issue
related to the CLI tooling for the lang server

The apollo deps also seem to require that we pin nodejs to 14.x (an LTS
which is still on maint through to April 2023), rather than the more
recent LTS.

Some small tweaks around tailwind may still be needed. With the CRA
update, we should no longer need the postcss cli since react-scripts
should now be running the postcss tasks for us as needed.

We'll see...