refactor: update Dockerfile to use musl and bundle
Updating the Dockerfule to produce a single static binary means we can
use the Dockerfile just to run the app as a docker image, but also as a
means to build and copy out the static binary.

The final stage image can now be `FROM scratch` meaning it will be very
very small. Currently, this is around 13Mb (down from around 96Mb).

In the past I've seen apps built with musl perform terribly. It was
rumored this was related to the default allocator employed by Rust.
Still, since this is a toy project, it's not a huge concern. YOLO.