feat: configure server via cli/env
Adds new CLI flags to configure:

- the bind address for the HTTP server
- the admin key (such that it can be something consistent rather than
  random)
- the deck type (formerly only configurable via env var)

An unfortunate inconsistency is in `PHI_STATIC_DIR` which remains only
configurable via env var (not included in the new set of CLI flags).

The reason for this is the asset baking happens at compile time and relies
on the same var, which also configures the location for SPA files on
disk in the non-baked mode.