refactor: switch to actix-web 4.x (from warp)
Whoops. After a brief look at warp, I really struggled to get an
understanding of the way the filters API works. The compiler feedback
was not useful in showing me what I was missing because of the huge
types (thanks to the combinators used to build filter chains).

After all that struggle, I've resigned to going back to actix-web. This
time, using the 4.x RC (seems like there might be a 4.x stable release
soon-ish?)