refactor: use cookies to track user id/name
This allows us to maintain continuity between play sessions. Effectively
this means a player id (auto-generated UUID) is "reserved" on the
client's machine once issued. If the player is removed from the game
session, a new entry will be made in its place when the client
reconnects (so long as the cookie is present).

Name changes are fiddly, requiring us to inspect the game state in the
graphql endpoint to decide if the cookie should be updated or not. This
is required as the actix-session `Session` object cannot be shared with
the graphql context (due to not being `Send`).
If things were different, a more desirable approach would be to have the
Mutation manipulate the cookie store itself when we see a name change.