refactor: use polling instead of websockets for game state updates
In order to keep the UI feeling responsive (since there's additional
latency involved when polling) several Mutation fields have been updated
to return a `Player`, allowing players to patch their own Apollo cache
when they initiate a change.

As for the disconnects, the threshold for disconnecting a player who's
heartbeat hasn't been seen has been extended to an hour (which is wildly
large). A 30 second threshold is used to vary the name tag for a player
so there's now a visual cue to show we haven't heard from a player
recently...

Combined with the changes in the earlier rev to leverage cookies, these
changes should mean player state is far more durable even with spotty
connectivity.